src/back-end/requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.38.0
django-guardian==1.4.9
django-rest-auth==0.9.3
djangorestframework==3.8.2
flake8==3.5.0
idna==2.7
Markdown==3.0.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
snowballstemmer==1.2.1
urllib3==1.23
